requirements.txt
Flask==3.0.3
ollama==0.3.1
openai==1.40.0
python-dotenv==1.0.1